Building a 2d grid and route movement
I focused on building out a 2d grid and then creating a function to
receive Left Right Up Down instructions.

With those instructions I basically log the route by assigning any
coordinates it crosses with a value of 1.

The tricky thing with using a list of lists, is that it seems to
basically reverse the logic. I am almost there but it is far
from logical to make sense.