Built route tracker and min intersection finder
I broke the task down into 2 objectives.

1) Track the route taken by a set of instructions
traverse_route() function takes in a set of instructions, interprets
them and then tracks in a list called "Traversed_coords" the coordinates
it passes along the way.

2) Get the overlapping coordinates, and find the closest value to origin
get_shortest_intersect() does this by taking in two respective outputs
or "routes" from traverse_route() and uses sets to create distinct
(unique) list of coordinates traversed, and then use the set.intersect
to find the overlapping coordinates. From there, I can find the smallest
sum of a tuple to get the closest to the origin.

At first this was returning a lower than expected value using the test
case. The problem I found was that if (L > R) or (D > U) overall, you'd
get negative coordinates tracked, and thus when you sum the tuples, the
coordinates with negative values are going to be overstated as the
shortest overlapping point. The simple fix was to use abs on the
tuple.